+++ core/testsuite/src/main/java/com/blazebit/persistence/testsuite/entity/PrimitiveFamily.java
@@ -1,0 +1,95 @@
+/*
+ * Copyright 2014 - 2022 Blazebit.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.blazebit.persistence.testsuite.entity;
+
+import javax.persistence.Basic;
+import javax.persistence.CascadeType;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
+import java.io.Serializable;
+
+/**
+ *
+ * @author Christian Beikov
+ * @since 1.6.8
+ */
+@Entity
+@Table(name = "prim_fam")
+public class PrimitiveFamily implements Serializable {
+    private static final long serialVersionUID = 1L;
+
+    private long id;
+    private String name;
+    private PrimitivePerson person;
+
+    public PrimitiveFamily() {
+    }
+
+    public PrimitiveFamily(String name) {
+        this.name = name;
+    }
+
+    @Id
+    @GeneratedValue
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    @Basic(optional = false)
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.REMOVE)
+    public PrimitivePerson getPerson() {
+        return person;
+    }
+
+    public void setPerson(PrimitivePerson person) {
+        this.person = person;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof PrimitiveFamily)) {
+            return false;
+        }
+
+        PrimitiveFamily that = (PrimitiveFamily) o;
+
+        return getId() == that.getId();
+    }
+
+    @Override
+    public int hashCode() {
+        return (int) (getId() ^ (getId() >>> 32));
+    }
+}